Crate dim_extern_api

Library contains a common interface for extracting and obtaining filename metadata as well as the implementations for various external APIs, such as TMDB.

Modules

filename

mock

tmdb

A TMDB client implementation with request coalescing and client-side rate-limiting.

Structs

ExternalActor

ExternalEpisode

ExternalMedia

ExternalSeason

Enums

Error

MediaSearchType

Traits

ExternalQuery

Trait that must be implemented by external metadata agents which allows the scanners to query for data.

ExternalQueryIntoShow

ExternalQueryShow

Trait must be implemented by all external metadata agents which support querying for tv shows.

IntoQueryShow

Type Aliases

Result