dim_extern_api

Module filename

Structs

Anitomy

An Anitomy parser instance.

CombinedExtractor

A special filename metadata extractor that combines torrent_name_parser and anitomy which in some cases is necessary. TNP is really good at extracting show titles but not season and episode numbers. Anitomy excels at this. Here we combine the title extracted by TPN and the season and episode number extracted by Anitomy.

Metadata

TorrentMetadata

Traits

FilenameMetadata