dim_extern_api

Module tmdb

A TMDB client implementation with request coalescing and client-side rate-limiting.

Structs

MetadataProviderOf

An instance of TMDBMetadataProvider with a generic parameter to infer the MediaType for searches.

Movies

Used to key MetadataProviderOf to search for movies, compliments TvShows.

TMDBMetadataProvider

TMDB Metadata Provider produces ExternalQuery implementors, and handles request coalescing and caching locally.

TvShows

Used to key MetadataProviderOf to search for TV shows, compliments Movies.

Constants

APP_USER_AGENT

The User-Agent is generated from the package name “dim” and the version so “dim/0.x.y.z”

TMDB_BASE_URL

The base url used to access TMDB;