Trait dim_extern_api::ExternalQueryShow

pub trait ExternalQueryShow: ExternalQuery {
    fn seasons_for_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        external_id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ExternalSeason>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn episodes_for_season<'life0, 'life1, 'async_trait>(
        &'life0 self,
        external_id: &'life1 str,
        season_number: u64
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ExternalEpisode>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

Trait must be implemented by all external metadata agents which support querying for tv shows.

Required Methods

fn seasons_for_id<'life0, 'life1, 'async_trait>(
&'life0 self,
external_id: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Vec<ExternalSeason>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Get all seasons for an external id. Seasons must be ranked in order by their number.

fn episodes_for_season<'life0, 'life1, 'async_trait>(
&'life0 self,
external_id: &'life1 str,
season_number: u64
) -> Pin<Box<dyn Future<Output = Result<Vec<ExternalEpisode>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Get all episodes for a season ranked in order of the episode number.